tutorials/partner-workshop-content/dsp_ml_flow_1/requirements.txt
alembic==1.13.1
aniso8601==9.0.1
asttokens==2.4.1
bayesian-optimization==1.4.3
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
Deprecated==1.2.14
docker==7.1.0
entrypoints==0.4
executing==2.0.1
Flask==3.0.3
fonttools==4.53.0
gitdb==4.0.11
GitPython==3.1.43
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
gunicorn==22.0.0
hana-ml==2.20.24042601
hdbcli==2.20.23
htmlmin==0.1.12
idna==3.7
importlib_metadata==7.1.0
ipython==8.25.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
Mako==1.3.5
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mlflow==2.13.2
numpy==1.26.4
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
plotly==5.22.0
prettytable==3.10.0
prompt_toolkit==3.0.47
protobuf==4.25.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.2
pydotplus==2.0.2
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
querystring-parser==1.2.4
requests==2.32.3
schedule==1.2.2
scikit-learn==1.5.0
scipy==1.13.1
shapely==2.0.4
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.30
sqlparse==0.5.0
stack-data==0.6.3
tenacity==8.3.0
threadpoolctl==3.5.0
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.16.0
zipp==3.19.2
